+++ .github/scripts/run_tests.sh
@@ -4,6 +4,8 @@
 
 set -eux -o pipefail
 
+sudo apt-get install wkhtmltopdf
+
 python -m venv "${HOME}/venv"
 source "${HOME}/venv/bin/activate"
 python -m site
@@ -15,6 +17,7 @@ python -m pip install xlrd
 python -m pip install dogpile.cache==0.9.2  # Later versions incompatible
 python -m pip install "SQLAlchemy<1.4"  # _get_immediate_cls_attr moved in 1.4
 python -m pip install pytest
+python -m pip install xhtml2pdf weasyprint pdfkit  # For PDF tests
 python -m pip install -e .
 
 pytest